==> notebooks/03_Gold_Layer_Tables_Creation.ipynb <==
%%sql

-- Fact_Loan
CREATE OR REPLACE TABLE Fact_Loan (
  LoanID STRING NOT NULL,
  CustomerID STRING NOT NULL,
  BranchID STRING NOT NULL,
  StartDate DATE NOT NULL,
  LoanAmount DOUBLE NOT NULL,
  TenureMonths INT NOT NULL,
  InterestRate DOUBLE,
  Status STRING,
  LoanType STRING,
  ClosureDate DATE,
  EarlyClosureFlag BOOLEAN
);

-- Fact_EMI
CREATE OR REPLACE TABLE Fact_EMI (
  EMIID STRING NOT NULL,
  LoanID STRING NOT NULL,
  EMINumber INT NOT NULL,
  DueDate DATE NOT NULL,
  EMIAmount DOUBLE NOT NULL,
  AmountPaid DOUBLE,
  PaymentDate DATE,
  Status STRING
);

-- Dim_Customer
CREATE OR REPLACE TABLE Dim_Customer (
  CustomerID STRING NOT NULL,
  CustomerName STRING,
  Gender STRING,
  DOB DATE,
  City STRING,
  PhoneNumber STRING,
  Email STRING,
  PAN STRING,
  KYCStatus STRING
);

-- Dim_Branch
CREATE OR REPLACE TABLE Dim_Branch (
  BranchID STRING NOT NULL,
  BranchName STRING,
  Region STRING,
  Manager STRING,
  ContactNumber STRING
);

-- Dim_RiskScore
CREATE OR REPLACE TABLE Dim_RiskScore (
  CustomerID STRING NOT NULL,
  RiskScore INT
);

-- Dim_LoanType
CREATE OR REPLACE TABLE Dim_LoanType (
  LoanType STRING NOT NULL, -- e.g., 2W, 4W, Tractor, Commercial, Used
  Category STRING 
);


df_loans = spark.read.parquet("Files/Silver/Loans/")
df_loans.write.mode("append").insertInto("Fact_Loan")

df_emis = spark.read.parquet("Files/Silver/EMIs/")
df_emis.write.mode("append").insertInto("Fact_EMI")

df_customers = spark.read.parquet("Files/Silver/Customers/")
df_customers.write.mode("append").insertInto("Dim_Customer")

df_branches = spark.read.parquet("Files/Silver/Branches/")
df_branches.write.mode("append").insertInto("Dim_Branch")

df_risks = spark.read.parquet("Files/Silver/RiskScore/")
df_risks.write.mode("append").insertInto("Dim_RiskScore")


%%sql
INSERT INTO Dim_LoanType VALUES
  ('2W', 'Two Wheeler'),
  ('4W', 'Four Wheeler'),
  ('Tractor', 'Tractors, Harvesters or Balers'),
  ('Commercial', 'Truck/Bus or Business Vehicles'),
  ('Used', 'Second-hand Vehicle');

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.fact_loan LIMIT 1000")
display(df.limit(10))

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.fact_emi LIMIT 1000")
display(df.limit(10))

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.dim_branch LIMIT 1000")
display(df.limit(10))

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.dim_customer LIMIT 1000")
display(df.limit(10))

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.dim_riskscore LIMIT 1000")
display(df.limit(10))

df = spark.sql("SELECT * FROM VehicleLoanLakehouse.dim_loantype LIMIT 1000")
display(df.limit(10))

%%sql
-- Create dim_Forecast table
CREATE OR REPLACE TABLE dim_Forecast (
    LoanID STRING NOT NULL,
    Forecast_OverdueAmount DOUBLE,
    Forecast_NPA_Percent DOUBLE,
    PredictedRiskScore INT,
    PriorityRecoveryFlag STRING
);



df_forecast = spark.read.parquet("Files/Silver/ForecastData/")
df_forecast.write.mode("append").insertInto("Dim_Forecast")
